# Statically resolve literal types at code-gen